Add files via upload
This is a cosine similarity algorithm I designed. It uses gensim's word2vec model trained on text8 to compare individual words in a document and through repetition and the use of thresholds the algorithm can effectively calculate the percent similarity between documents. This is a novel approach to Natural Language Processing. The only downside of this algorithm is, however, that it does not efficiently compute the percent similarity in a large dataset. This algorithm was used to convert the proprietary lab tests of ResultCare to the universal standard LOINC.